webgpu: Add a `webgpu_traits` crate
This breaks the dependency on `script_traits` on
`webgpu`. In general, the `traits` crates shouldn't depend on Servo
non-`traits` crates. This is necessary to move "script to constellation"
messages to the `constellation_traits` crate, making it the entire API
for talking to the constellation. This will break a circular dependency
when that happens.
