+++ Cargo.toml
@@ -162,6 +162,7 @@ unicode-segmentation = "1.12.0"
 url = "2.5"
 uuid = { version = "1.12.1", features = ["v4"] }
 webdriver = "0.51.0"
+webgpu_traits = { path = "components/shared/webgpu" }
 webpki-roots = "0.26"
 webrender = { git = "https://github.com/servo/webrender", branch = "0.66", features = ["capture"] }
 webrender_api = { git = "https://github.com/servo/webrender", branch = "0.66" }
