+++ components/script/dom/webgpu/gpubindgroup.rs
@@ -6,7 +6,7 @@ use std::borrow::Cow;
 
 use dom_struct::dom_struct;
 use webgpu::wgc::binding_model::BindGroupDescriptor;
-use webgpu::{WebGPU, WebGPUBindGroup, WebGPUDevice, WebGPURequest};
+use webgpu_traits::{WebGPU, WebGPUBindGroup, WebGPUDevice, WebGPURequest};
 
 use crate::conversions::Convert;
 use crate::dom::bindings::cell::DomRefCell;
