+++ components/script/dom/webgpu/gpucompilationinfo.rs
@@ -4,7 +4,7 @@
 
 use dom_struct::dom_struct;
 use js::rust::MutableHandleValue;
-use webgpu::ShaderCompilationInfo;
+use webgpu_traits::ShaderCompilationInfo;
 
 use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUCompilationInfoMethods;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object_with_proto};
