+++ components/shared/webgpu/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "webgpu_traits"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+publish.workspace = true
+rust-version.workspace = true
+
+[lib]
+name = "webgpu_traits"
+path = "lib.rs"
+
+[dependencies]
+arrayvec = { workspace = true }
+base = { workspace = true }
+ipc-channel = { workspace = true }
+malloc_size_of = { workspace = true }
+serde = { workspace = true }
+webrender_api = { workspace = true }
+wgpu-core = { workspace = true }
+wgpu-types = { workspace = true }
