+++ components/shared/webgpu/messages/mod.rs
@@ -5,3 +5,7 @@
 pub mod recv;
 pub mod to_dom;
 pub mod to_script;
+
+pub use recv::*;
+pub use to_dom::*;
+pub use to_script::*;
